Imported current western code & setup config directory with builder.py changes
This is the current version of the western code that we will use as the most up-to-date NVCS files. I also setup the config directory where we will branch based on East, West, and Alaska

Since we're changing the nvcs_config file to contain content from each region, the builder file needed to be updated to accommodate this. Hopefully it should be easier to quickly swap these settings by simplying updating the projectRoot and uncommenting the region line we want.