Implemented DebugConfig class
This class will rely on an .ini file to store paths to easily swap between testing the western, eastern, and Alaska keys. This means to test things like builder.py & tester.py, we just need to update the debug_config.ini file with the correct paths and update the "TargetConfig" value as needed. Hopefully this should make swapping between the classification keys a lot easier. Unfortunately, there are still a few hard-coded paths when it comes to importing in the tester.py file that I want to eventually resolve.