Adjusted Forest Plantations filter/trigger logic
The Forest Plantation filter/triggers were unconventional compared to the rest of the nodes. It had two entries under its PLANTED_EXOTIC_OR_POPULOUS filter which both set planted to "yes" and then either species to "Populus" or exotic to "yes". Afterwards, its RIV was calculated on the whole filter to see if it exceeded zero. To make this more similar to other nodes, I separated those two into separate filter categories and used an or statement to check if PLANTED_EXOTIC > 0 or PLANTED_POPULUS > 0 after first matching the existing PLANTATION requirement. This change enforced consistency and even lowered the unclassified counts by double digits for 2007, 2013, and 2017 inventory year tests.