Fixed search functionality
This broke when I added duplicate buttons into the warning lists. This is because I was using the ID property for shifting focus instead of using a data attribute. Now I can narrow down focus to just the buttons within the tree rather than accidentally focusing the ones in the warning lists higher up the page.