Began implementing some package logic
As this app gets closer to a packaged & deployed state, I had to revisit some of our relative paths and make sure they're correctly setup. We'll copy the NVCS western key nodes into the deployed project and make sure to path to it for things like JSON retrievals and updates. Also, I standardized on an "NVCS Configuration Tool" name for now.