Added load packaged JSON to differentiate custom JSON loads & improved error handling
Now we'll be able to load the JSON packaged with the application OR custom JSON that was stored somewhere else. Paths were cleaned up to be absolute during final saves to get rid of any weird relative path logging. I also dropped the try/catchs around the actual handling index.js functions and instead relied on the frontend to communicate that an error occurred