Fixed bug in M298
The description calls for ruderal or exotic species having an RIV greater than or equal to 80, yet the filter name only said EXOTIC and it only had the exotic filter, causing the key to have an error for that macrogroup logic. I fixed this up to match the description present in the 17.1 revisions.