Adjusted D338 & M024
It turns out that M024 was basically just D338 minus an exotic species inclusion. So I simply took M024's description and appended it to D338's and adjusted the triggers and filters to match. Then M024 became a "true" trigger because its the only macrogroup as stated by the 17.1 revisions.