Fixed default config file from being overwritten
Because the Python classifier reads from the debug_config.ini file in the nvcs-dev directory, we need to update that file with custom settings that the user provides, such as inventory years or DB paths. However, that also means default settings get overwritten whenever the tester script is run. To accommodate this, I made sure to clone the debug_config.ini file and use it for retrieving default setting. For packaged applications, this is automatically copied over using the extraResource option in forge.config.js. For non-packaged applications, it will be created at runtime on application start.