Improved accessibility of all dialogs
We now make use of the <dialog> tags supported in HTML and the close and open events that exist for them. Overlays are now the backdrop CSS element, so I was able to remove all of that logic, too. Finally, the testing overlay is now a testing dialog that we make appear using similar conventions. Now keyboard focus and focus order should be working as expected.