Fixed validations for ROOT element
The ROOT really only should be updated to change its direct child node ordering. Therefore, I updated all the validation logic to only occur when not working with the ROOT element. I also updated the JSON dialog to set all of the fields and buttons not applicable to readonly or disabled when working within the context of the ROOT element. Finally, I removed a redundant class definition in our CSS file.