Resolved duplicate button description accessibility errors
Now the error/warning-duplicated hierarchy buttons will be given an aria-label to identify their violation. This should eliminate multiple buttons with the same ANDI output. I also fixed a bug in the error alerts where I was passing "errors" instead of "error" and causing the warning message to appear. This shouldn't occur again because of new conditional logic to throw an error if "error" or "warning" isn't passed. The nested warning buttons are now also aria-described-by the lists they control for a unique ANDI output.