Began implementing ClassificationKey.java
This is going to be a template eventually, but I wanted to set up some examples of the first nodes being added so that I can replicate that generation behavior later. I also had to create an interface for mapping out the match function and utilizing it from within the Node class.