Implemented testing for the Java classification system
This still needs to have some big improvements for performance, but we can now query all of the key test data into JSON format and feed it into the Java classifier by executing it through the command line in the tester_java.py script. I had to make some adjustments so that we'd return full paths for better logging purposes and also handle null strings better. Again, this will need to be improved on for performance because it takes too long when working with thousands of conditions.