Improved performance of Java classifier testing
Now we'll write the entire IDENT and ROW_JSON query to a temporary file location and read it in directly from the Java, meaning we now pass input and output paths into the App.java file as opposed to raw JSON. The Java script will then go line by line and classify out each plot to its own .out file. For the sake of line-by-line comparison, I also adjusted the plot.io script to order by IDENT when querying for rows in the Python classification tester.