Fixed Java PatternLists when working with multiple pairs per PatternGroup
There are some situations where the key relies on two patterns to match together in order to pass, such as EXOTIC_TREE_SPP in the generated element_053() function. To fix this, we'll now populate the PatternList with a list of PatternGroups which is just a list of Patterns (pairs). Wrapping this previous list with another list leads to more generated code, but now we match Python and support AND conditionals on a match requirement instead of just OR conditionals.