Moved null species exclusions from tester to riv/spcov calculation
I realized we shouldn't be skipping null species in the tester because when it comes to production implementation, we'll need to ensure those were being skipped, too. Now we'll avoid skipping in the tester and instead skip any null species when calculating riv/spcov percentages. This change, when run against all the input data, showed no impact on results. I also re-built the eastern Python key because it looks like I forgot to do that in an earlier pull request.