Added ability to generate flattened alerts
We'll now call createFlattenedAlerts() which will parse the errors and warnings arrays, take each individual violation, and report them in a flattened array to eventually be used in a database table. This required each individual error/warning to be given a unique alert ID. I also had to refactor some of these existing error/warning structures to be more consistent and allow the use of these unique alert IDs.