Fixed bug with string values containing zeroes
When querying from a field like AFFORESTATION_CD which can contain zeroes, but the value is casted to a string for categorical purposes, Python would choose an empty string over it due to how zeroes are parsed in boolean logic. To fix this, we'll check to see if the value is None before using the alternative on all data fields.