Temporary fix for Y/N RUDERAL and EXOTIC values
The eastern data has RUDERAL and EXOTIC values as Y/N as opposed to yes/no. This temporary solution will update the input view to cast those values to the appropriate format. We may need to revert this down the line if the values end up getting adjusted in future SQLite input database updates.